Rework how pkg/signature/payload deals with identities
The principal purpose of the critical.image.docker-reference fields is to
_exactly_ identify the _purpose_ the signer claims the image to fulfill.
It's expressed as a registry/.../repo:tag , because that's how users deal
with image purpose and identity, and users inherently provide that value
(without needing to add a whole new "purpose" field to everything in the ecosystem).

So far, cosign has only been setting identity to a repo (which is insufficient
and lets registries to substitute diferent tags = versions of the same application),
and payload.Cosign was not even allowing to create a tagged identity.

So, explicitly separate Cosign.ClaimedIdentity from Cosign.ImageDigest,
so that users can set both independently (and are forced to think about the claimed
identity a bit more).

Update pkg/signature.SignImage (which is not used by cosign, but
has one external user on GitHub) correspondingly.

NOTE: This breaks existing creators of pkg/signature/payload.Cosign and
pkg/signature.SignImage; that's intentional, the current Image field does not
contain enough information to create a good signature.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>